+++ freeman-core/src/Core/Plugin.php
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Plugin {
 
-	const VERSION = '1.20.1';
+	const VERSION = '1.21.0';
 
 	/**
 	 * Singleton instance.